分离单任务模型和多任务模型
